backend/requirements.txt
fastapi
uvicorn
openai
langchain-openai
python-dotenv
faiss-cpu
PyPDF2
python-docx
chromadb
langchain>=0.1.14
langchain-community
websockets
python-multipart


